requirements.txt
torch
torchvision
numpy
scikit-learn
efficientnet-pytorch
Pillow
